requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
requests>=2.31.0
python-pptx>=0.6.21
pillow>=10.0.0
openai>=1.0.0
